pyproject.toml
[project]
name = "stream-demos-datagen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiokafka>=0.12.0",
    "pandas>=2.2.3",
    "pygame>=2.6.1",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
]

[dependency-groups]
dev = [
    "ansible>=11.2.0",
    "ansible-lint>=25.1.1",
    "boto3>=1.36.12",
    "botocore>=1.36.12",
    "docker-squash>=1.2.2",
    "ipykernel>=6.29.5",
    "pyiceberg>=0.8.1",
    "watchdog>=6.0.0",
]
